# backend/src/main/resources/openapi/API_REST.yaml
openapi: 3.0.3
info:
  title: Simulador MVP API
  version: 1.0.0
  description: |
    API REST del Simulador — plataforma de simulación empresarial competitiva paraguaya.

    ## Autenticación
    Todas las rutas excepto `/auth/login` y `/auth/refresh` requieren `Authorization: Bearer {access_token}`.
    El access token expira en **15 minutos**. Usar `POST /auth/refresh` con el refresh token (válido **30 días**) para obtener uno nuevo.

    ## Autorización
    Tres roles: `MODERADOR`, `JUGADOR`, `ADMIN_PLATAFORMA`.
    - `MODERADOR`: gestiona competencias que creó.
    - `JUGADOR`: participa en equipos, registra decisiones.
    - `ADMIN_PLATAFORMA`: gestiona catálogos (rubros, parámetros, eventos).

    ## Paginación
    Endpoints de listado aceptan `page` (default 0) y `size` (default 20, max 100). Responden con metadata.

    ## Errores
    Formato uniforme (RFC 7807 Problem Details) con `status`, `type`, `title`, `detail`, `errors[]` (campo a campo).

    ## Moneda
    Todos los montos se expresan en **guaraníes como enteros** (sin centavos).

    ## Tiempo real
    WebSocket en `/ws/competencias/{codigo}` para push de eventos (trimestre cerrado, evento disparado, resultados listos).

  contact:
    name: Equipo Simulador
    email: dev@simulador.py
servers:
  - url: https://api.simulador.py/v1
    description: Producción
  - url: https://api-stg.simulador.py/v1
    description: Staging
  - url: http://localhost:8080/v1
    description: Local

...
paths:
...
  /usuarios:
    get:
      tags: [Usuarios]
      summary: Listar usuarios (ADMIN)
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: rol
          in: query
          schema:
            type: string
            enum: [MODERADOR, JUGADOR, ADMIN_PLATAFORMA]
        - name: activo
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          description: Búsqueda por email o nombre
          schema:
            type: string
      responses:
        '200':
          description: Lista paginada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedUsuarios'
    post:
      tags: [Usuarios]
      summary: Crear usuario (ADMIN o MODERADOR para invitar jugadores)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioCreate'
      responses:
        '201':
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '422':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/Conflict'
  /usuarios/{id}:
    get:
      tags: [Usuarios]
      summary: Obtener usuario
      parameters:
        - $ref: '#/components/parameters/UsuarioId'
      responses:
        '200':
          description: Usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Usuarios]
      summary: Actualizar usuario (propio o ADMIN)
      parameters:
        - $ref: '#/components/parameters/UsuarioId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioUpdate'
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
    delete:
      tags: [Usuarios]
      summary: Desactivar usuario (soft-delete, solo ADMIN)
      parameters:
        - $ref: '#/components/parameters/UsuarioId'
      responses:
        '204':
          description: Desactivado

  # ------------------------------------------------------------------
  # CATÁLOGOS
  # ------------------------------------------------------------------
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ----- PARAMS -----
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    UsuarioId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Conflicto de estado (ej. transición inválida, unicidad)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ValidationError:
      description: Error de validación de entrada
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemValidacion'

  # ----- SCHEMAS -----
  schemas:
...
    Problem:
      type: object
      properties:
        type:
          type: string
          example: 'https://simulador.py/errors/validation'
        title:
          type: string
          example: 'Validación fallida'
        status:
          type: integer
          example: 422
        detail:
          type: string
        instance:
          type: string
          example: '/competencias/123'
    ProblemValidacion:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: precio_venta
                  message:
                    type: string
                    example: 'Debe ser mayor a 0'

    # --- Auth ---
...
    Usuario:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        nombre_completo:
          type: string
        rol:
          type: string
          enum: [MODERADOR, JUGADOR, ADMIN_PLATAFORMA]
        activo:
          type: boolean
        email_verificado:
          type: boolean
        created_at:
          type: string
          format: date-time
        last_login_at:
          type: string
          format: date-time
          nullable: true
    UsuarioCreate:
      type: object
      required: [email, password, nombre_completo, rol_codigo]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        nombre_completo:
          type: string
          maxLength: 150
        rol_codigo:
          type: string
          enum: [MODERADOR, JUGADOR, ADMIN_PLATAFORMA]
    UsuarioUpdate:
      type: object
      properties:
        nombre_completo:
          type: string
        password:
          type: string
          format: password
          minLength: 8
        rol_codigo:
          type: string
          enum: [MODERADOR, JUGADOR, ADMIN_PLATAFORMA]
    PagedUsuarios:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Usuario'
        page:
          type: integer
        size:
          type: integer
        total_elements:
          type: integer
          format: int64
        total_pages:
          type: integer

    # --- Rubro ---
...
security:
  - bearerAuth: []